Narrow scope of ubuntu-server check to xenial release
In order to avoid increasing the the limit across all releases, past,
present, and future, will increase the limit on Xenial only. Xenial is
currently the only release that is across the board over the 700Mb limit
and all others still fall under that value. This will allow to continue
watching the development releases grow.