livefs: Determine the OEM flavouor to use dynamically
The idea is that we include these kernels

  LTS xx.yy: generic oem-xx.yy
  LTS+1: generic oem-xx.yy
  LTS+2: generic oem-xx.yy
  LTS+3: generic oem-xx.yy
  LTS aa.bb: generic oem-aa.bb
  LTS+1: generic oem-aa.bb

and so on, assuming that the OEM kernel only gets its name revved in LTS cycles

This is all subject to further iteration. Here we're trying to get things into
shape so that this take care of itself for future release cycles.